+++ x/stake/keeper_keys.go
@@ -5,6 +5,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/wire"
+	crypto "github.com/tendermint/go-crypto"
 )
 
 // TODO remove some of these prefixes once have working multistore
@@ -18,12 +19,9 @@ var (
 	ValidatorsKey          = []byte{0x03} // prefix for each key to a validator
 	AccUpdateValidatorsKey = []byte{0x04} // prefix for each key to a validator which is being updated
 	RecentValidatorsKey    = []byte{0x05} // prefix for each key to the last updated validator group
-
 	ToKickOutValidatorsKey = []byte{0x06} // prefix for each key to the last updated validator group
-
 	DelegatorBondKeyPrefix = []byte{0x07} // prefix for each key to a delegator's bond
-
-	CounterKey = []byte{0x08} // key for block-local tx index
+	CounterKey             = []byte{0x08} // key for block-local tx index
 )
 
 const maxDigitsForAccount = 12 // ~220,000,000 atoms created at launch
@@ -48,12 +46,13 @@ func GetAccUpdateValidatorKey(addr sdk.Address) []byte {
 	return append(AccUpdateValidatorsKey, addr.Bytes()...)
 }
 
-// get the key for the accumulated update validators
-func GetRecentValidatorKey(addr sdk.Address) []byte {
+// get the key for the recent validator group, ordered like tendermint
+func GetRecentValidatorKey(pk crypto.PubKey) []byte {
+	addr := pk.Address()
 	return append(RecentValidatorsKey, addr.Bytes()...)
 }
 
-// reverse operation of GetRecentValidatorKey
+// remove the prefix byte from a key
 func AddrFromKey(key []byte) sdk.Address {
 	return key[1:]
 }